// frontend/src/components/ui/button.d.ts
import * as React from "react";
export declare const Button: React.ForwardRefExoticComponent<React.ButtonHTMLAttributes<HTMLButtonElement> & {
    variant?: "default" | "outline" | "ghost" | "destructive";
    size?: "sm" | "md" | "lg" | "icon";
} & React.RefAttributes<HTMLButtonElement>>;
export default Button;